requirements.txt
zulip==0.7.1
requests==2.24.0
python3-logstash==0.4.80